Implement reading calibration events
