+++ ctapipe/io/tests/test_simteleventsource.py
@@ -8,6 +8,7 @@ from itertools import zip_longest
 
 gamma_test_large_path = get_dataset_path("gamma_test_large.simtel.gz")
 gamma_test_path = get_dataset_path("gamma_test.simtel.gz")
+calib_events_path = get_dataset_path('calib_events.simtel.gz')
 
 
 def compare_sources(input_url):
@@ -161,3 +162,41 @@ def test_additional_meta_data_from_mc_header():
             value.to_value(expectation.unit),
             expectation.to_value(expectation.unit)
         )
+
+
+def test_hessio_file_reader():
+    dataset = gamma_test_path
+    with SimTelEventSource(input_url=dataset) as reader:
+        assert reader.is_compatible(dataset)
+        assert not reader.is_stream
+        for event in reader:
+            if event.count == 0:
+                assert event.r0.tels_with_data == {38, 47}
+            elif event.count == 1:
+                assert event.r0.tels_with_data == {11, 21, 24, 26, 61, 63, 118,
+                                                   119}
+            else:
+                break
+        for event in reader:
+            # Check generator has restarted from beginning
+            assert event.count == 0
+            break
+
+    # test that max_events works:
+    max_events = 5
+    with SimTelEventSource(input_url=dataset, max_events=max_events) as reader:
+        count = 0
+        for _ in reader:
+            count += 1
+        assert count == max_events
+
+    # test that the allowed_tels mask works:
+    with SimTelEventSource(input_url=dataset, allowed_tels={3, 4}) as reader:
+        for event in reader:
+            assert event.r0.tels_with_data.issubset(reader.allowed_tels)
+
+
+def test_calibration_events():
+    with SimTelEventSource(input_url=calib_events_path, skip_calibration=False) as reader:
+        for e in reader:
+            pass
