+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - default
   - cta-observatory
 dependencies:
-  - ctapipe-extra=0.2.16
+  - ctapipe-extra=0.2.17
   - zeromq
   - pyzmq>=17  # needed for correct function of notebooks on OSX
   - astropy
